refactor: remove legacy types and consolidate to feature-based structure
