+++ src/types/index.ts
@@ -5,6 +5,3 @@ export * from "@/features/tenants/types";
 export * from "@/features/auth/types";
 export * from "@/features/providers/types";
 export * from "@/features/shared";
-
-// legacy exports for backward compatibility
-// Note: form-data types have been moved to their respective features
