--- src/types/tenant.ts
@@ -1,13 +1,0 @@
-import { UtilityProviderCategory } from "@/types/utility-provider";
-
-export interface Tenant {
-	readonly id: string;
-	readonly userId: string;
-	readonly name: string;
-	readonly email: string;
-	readonly secondaryName?: string | undefined;
-	readonly shares: {
-		readonly [key in UtilityProviderCategory]: number;
-	};
-	readonly outstandingBalance: number;
-}
